remove leftovers and tidy up non-functional codes
