--- src/main/java/seedu/address/ui/TaskCard.java
@@ -1,65 +1,0 @@
-package seedu.address.ui;
-
-import javafx.fxml.FXML;
-import javafx.scene.Node;
-import javafx.scene.control.Label;
-import javafx.scene.layout.HBox;
-import seedu.address.model.task.ReadOnlyTask;
-
-public class TaskCard extends UiPart{
-
-    private static final String FXML = "TaskListCard.fxml";
-
-    @FXML
-    private HBox cardPane;
-    @FXML
-    private Label name;
-    @FXML
-    private Label id;
-    @FXML
-    private Label phone;
-    @FXML
-    private Label address;
-    @FXML
-    private Label email;
-    @FXML
-    private Label tags;
-
-    private ReadOnlyTask task;
-    private int displayedIndex;
-
-    public TaskCard(){
-
-    }
-
-    public static TaskCard load(ReadOnlyTask task, int displayedIndex){
-        TaskCard card = new TaskCard();
-        card.task = task;
-        card.displayedIndex = displayedIndex;
-        return UiPartLoader.loadUiPart(card);
-    }
-
-    @FXML
-    public void initialize() {
-//        name.setText(task.getName().fullName);
-//        id.setText(displayedIndex + ". ");
-//        phone.setText(task.getPhone().value);
-//        address.setText(task.getAddress().value);
-//        email.setText(task.getEmail().value);
-//        tags.setText(task.tagsString());
-    }
-
-    public HBox getLayout() {
-        return cardPane;
-    }
-
-    @Override
-    public void setNode(Node node) {
-        cardPane = (HBox)node;
-    }
-
-    @Override
-    public String getFxmlPath() {
-        return FXML;
-    }
-}
