--- src/main/resources/view/DefaultBrowserPlaceHolderScreen.fxml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.text.Font?>
-<HBox xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" AnchorPane.bottomAnchor="0.0"
-      AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
-    <children>
-        <Label alignment="CENTER" text="To view contact's web page, click on the contact on the left."
-               textAlignment="CENTER" textFill="#a69a9a" wrapText="true" HBox.hgrow="ALWAYS">
-            <font>
-                <Font size="30.0"/>
-            </font>
-            <padding>
-                <Insets bottom="10.0" left="10.0" right="10.0" top="10.0"/>
-            </padding>
-        </Label>
-    </children>
-</HBox>
-
