+++ src/main/resources/view/NewMainWindow.fxml
@@ -11,7 +11,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.VBox?>
 
-<VBox prefHeight="547.0" prefWidth="824.0" xmlns="http://javafx.com/javafx/8.0.65" xmlns:fx="http://javafx.com/fxml/1" fx:controller="seedu.address.ui.NewMainWindow">
+<VBox prefHeight="547.0" prefWidth="824.0" xmlns="http://javafx.com/javafx/8.0.65" xmlns:fx="http://javafx.com/fxml/1" fx:controller="seedu.address.ui.MainWindow">
    <children>
       <MenuBar VBox.vgrow="NEVER">
          <menus>
